update structs and events - use uint32 instead of uint24 for reward epoch ids

+++ contracts/protocol/implementation/FlareSystemsCalculator.sol
@@ -92,7 +92,7 @@ contract FlareSystemsCalculator is Governed, AddressUpdatable, IIFlareSystemsCal
      */
     function calculateRegistrationWeight(
         address _voter,
-        uint24 _rewardEpochId,
+        uint32 _rewardEpochId,
         uint256 _votePowerBlockNumber
     )
         external onlyVoterRegistry
