+++ contracts/userInterfaces/IFlareSystemsCalculator.sol
@@ -9,7 +9,7 @@ interface IFlareSystemsCalculator {
     /// Event emitted when the registration weight of a voter is calculated.
     event VoterRegistrationInfo(
         address indexed voter,
-        uint24 indexed rewardEpochId,
+        uint32 indexed rewardEpochId,
         address delegationAddress,
         uint16 delegationFeeBIPS,
         uint256 wNatWeight,
